requirements.txt
Django>=4.2.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.2.0
matplotlib>=3.7.0
seaborn>=0.12.0
python-dotenv>=1.0.0
django-crispy-forms>=2.0
crispy-bootstrap5>=0.7
Pillow>=10.0.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.0 
djangorestframework-simplejwt>=5.2.2 